Fix tests. PR feedback
