+++ python/packages/ag-ui/examples/agents/document_writer_agent.py
@@ -4,6 +4,7 @@
 
 from agent_framework import ChatAgent, ai_function
 from agent_framework.azure import AzureOpenAIChatClient
+
 from agent_framework_ag_ui import AgentFrameworkAgent, DocumentWriterConfirmationStrategy
