+++ python/packages/ag-ui/examples/agents/recipe_agent.py
@@ -6,8 +6,9 @@ from enum import Enum
 
 from agent_framework import ChatAgent, ai_function
 from agent_framework.azure import AzureOpenAIChatClient
-from agent_framework_ag_ui import AgentFrameworkAgent, RecipeConfirmationStrategy
 from pydantic import BaseModel, Field
+
+from agent_framework_ag_ui import AgentFrameworkAgent, RecipeConfirmationStrategy
 
 
 class SkillLevel(str, Enum):
