+++ python/packages/ag-ui/examples/agents/research_assistant_agent.py
@@ -6,6 +6,7 @@ import asyncio
 
 from agent_framework import ChatAgent, ai_function
 from agent_framework.azure import AzureOpenAIChatClient
+
 from agent_framework_ag_ui import AgentFrameworkAgent
