+++ python/packages/ag-ui/examples/agents/task_planner_agent.py
@@ -4,6 +4,7 @@
 
 from agent_framework import ChatAgent, ai_function
 from agent_framework.azure import AzureOpenAIChatClient
+
 from agent_framework_ag_ui import AgentFrameworkAgent, TaskPlannerConfirmationStrategy
