+++ python/packages/ag-ui/examples/agents/ui_generator_agent.py
@@ -6,6 +6,7 @@ from typing import Any
 
 from agent_framework import ChatAgent, ai_function
 from agent_framework.azure import AzureOpenAIChatClient
+
 from agent_framework_ag_ui import AgentFrameworkAgent
